+++ x/evm/types/utils.go
@@ -8,6 +8,10 @@ import (
 	"github.com/pkg/errors"
 	"golang.org/x/crypto/sha3"
 
+	abci "github.com/tendermint/tendermint/abci/types"
+	tmtypes "github.com/tendermint/tendermint/types"
+	"github.com/tendermint/tendermint/version"
+
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
@@ -151,4 +155,37 @@ func recoverEthSig(R, S, Vb *big.Int, sigHash ethcmn.Hash) (ethcmn.Address, erro
 	copy(addr[:], ethcrypto.Keccak256(pub[1:])[12:])
 
 	return addr, nil
+}
+
+// AbciHeaderToTendermint is a util function to parse a tendermint ABCI Header to
+// tendermint types Header.
+func AbciHeaderToTendermint(header abci.Header) tmtypes.Header {
+	return tmtypes.Header{
+		Version: version.Consensus{
+			Block: version.Protocol(header.Version.Block),
+			App:   version.Protocol(header.Version.App),
+		},
+		ChainID: header.ChainID,
+		Height:  header.Height,
+		Time:    header.Time,
+
+		LastBlockID: tmtypes.BlockID{
+			Hash: header.LastBlockId.Hash,
+			PartsHeader: tmtypes.PartSetHeader{
+				Total: int(header.LastBlockId.PartsHeader.Total),
+				Hash:  header.LastBlockId.PartsHeader.Hash,
+			},
+		},
+		LastCommitHash: header.LastCommitHash,
+		DataHash:       header.DataHash,
+
+		ValidatorsHash:     header.ValidatorsHash,
+		NextValidatorsHash: header.NextValidatorsHash,
+		ConsensusHash:      header.ConsensusHash,
+		AppHash:            header.AppHash,
+		LastResultsHash:    header.LastResultsHash,
+
+		EvidenceHash:    header.EvidenceHash,
+		ProposerAddress: header.ProposerAddress,
+	}
 }